Update PhakeRamayanaUnicode.otf
Two versions of letter NYA are included. By default the 107A now has the second bowl open on top, like the Shan letter 107A. The version with the second bowl open on the bottom is now encoded as A9E7, the Tai Laing version of the letter.

Not sure if it makes sense, please let's try it out.